requirements.txt
aiogram==3.x
aiohttp==3.8.6
python-dotenv==1.0.0
